5ZCMF7C split epiq-api.ts into domain modules under mcp/api
