+++ source/mcp/epiq-api.ts
@@ -1,2747 +1,45 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import {ulid} from 'ulid';
-import {getStateBranchRoot} from '../git/git-storage.js';
-import {execGit} from '../git/git-utils.js';
-import {getStateBranch} from '../git/git-constants.js';
-import {ensureLocalStateBranch, ensureStateBranchWorktree} from '../git/git.js';
-import {syncEpiqWithRemote} from '../git/sync.js';
-import {applyActorNameArgument} from '../lib/config/actor-env.js';
-import {loadSettingsFromConfig} from '../lib/config/user-config.js';
-import {createIssueEvents} from '../lib/event/common-events.js';
-import {ulidTimeMs} from '../lib/event/date-utils.js';
-import {bootStateFromEventLog} from '../lib/event/event-boot.js';
-import {
-	accountedSignature,
-	accountFor,
-	logSignature,
-} from '../lib/event/log-signature.js';
-import {getEpiqDirPath} from '../lib/storage/paths.js';
-import {
-	loadEventActors,
-	loadMergedEvents,
-	loadMergedEventsWithUnreadable,
-} from '../lib/event/event-load.js';
-import {materializeAndPersistAll} from '../lib/event/event-materialize-and-persist.js';
-import {getPersistFileName} from '../lib/event/event-persist.js';
-import {AppEvent, MovePosition} from '../lib/event/event.model.js';
-import {filterEventsForBoard} from './timeline-index.js';
-import {resolveReopenParentFromLog} from '../lib/event/log-utils.js';
-import {CLOSED_SWIMLANE_ID} from '../lib/event/static-ids.js';
-import {
-	isBoardNode,
-	isSwimlaneNode,
-	isTicketNode,
-	Swimlane,
-	Ticket,
-} from '../lib/model/context.model.js';
-import {failed, isFail, Result, succeeded} from '../lib/model/result-types.js';
-import {nodeRepo} from '../lib/repository/node-repo.js';
-import {
-	resolveAndPersistRankForCreate,
-	resolveAndPersistRankForMove,
-} from '../lib/repository/rank.js';
-import {getSafeState, isStateInitialized} from '../lib/state/state.js';
-import {setSynced, setSyncFailed, setSyncing} from '../lib/state/sync-state.js';
-import {recordRecentProject} from '../lib/config/recent-projects.js';
-import {resolveClosestEpiqProjectRoot} from '../lib/storage/paths.js';
-import {
-	Contributor,
-	REMOVED_CONTRIBUTOR_NAME,
-} from '../lib/model/app-state.model.js';
-import {describeEvent} from '../lib/event/format-log-utils.js';
-import {getStringColor} from '../lib/utils/color.js';
-import {
-	MAX_ASSIGNEE_NAME_LENGTH,
-	MAX_ASSIGNEES_PER_CREATE,
-	MAX_COMMENT_LENGTH,
-	MAX_DESCRIPTION_LENGTH,
-	MAX_TAG_NAME_LENGTH,
-	MAX_TAGS_PER_CREATE,
-	MAX_TITLE_LENGTH,
-	tooLong,
-} from '../lib/utils/text.limits.js';
-import {nodeRef, nodeRefMatches} from '../lib/utils/node-ref.js';
-import {sanitizeInlineText} from '../lib/utils/string.utils.js';
-import {
-	DEFAULT_ATTACHMENT_MAX_KB,
-	getAttachmentFileName,
-	getAttachmentMarkdown,
-	resolveAttachmentBlob,
-	writeAttachmentBlob,
-} from '../lib/media/media-store.js';
-import {logger} from '../logger.js';
-import {
-	ApiAssignee,
-	ApiBatchOutcome,
-	ApiIssue,
-	ApiIssueBrief,
-	ApiIssueComment,
-	ApiIssueDetail,
-	ApiState,
-	ApiIssueHistoryEntry,
-	ApiSwimlane,
-} from './api-state.model.js';
-import {getTimeTravelStatus} from './epiq-time-travel.js';
-
-type ToolInput = {
-	repoRoot?: string;
-};
-
-type SyncInput = ToolInput;
-
-// One ticket or many. issueIds asks for a per-ticket outcome; issueId alone
-// keeps the single answer.
-type IssueTargets = {
-	issueId?: string;
-	issueIds?: string[];
-};
-
-type MoveIssueInput = ToolInput &
-	IssueTargets & {
-		parentId: string;
-		position?: MovePosition;
-	};
-
-type ListIssuesInput = ToolInput & {
-	includeClosed?: boolean;
-	boardId?: string;
-	swimlaneId?: string;
-	tag?: string;
-	assignee?: string;
-	query?: string;
-	brief?: boolean;
-};
-
-type ListSwimlanesInput = ToolInput & {
-	boardId?: string;
-};
-
-type CreateSwimlaneInput = ToolInput & {
-	title: string;
-	boardId: string;
-};
-
-type EditSwimlaneTitleInput = ToolInput & {
-	swimlaneId: string;
-	title: string;
-};
-
-type MoveSwimlaneInput = ToolInput & {
-	swimlaneId: string;
-	boardId: string;
-	position?: MovePosition;
-};
-
-type DeleteSwimlaneInput = ToolInput & {
-	swimlaneId: string;
-};
-
-type CreateIssueInput = ToolInput & {
-	title: string;
-	parentId: string;
-	description?: string;
-	tagNames?: string[];
-	assigneeNames?: string[];
-};
-
-type CloseIssueInput = ToolInput & IssueTargets;
-
-type BootResult = {
-	repoRoot: string;
-	stateBranchRoot: string;
-};
-
-type Actor = {
-	userId: string;
-	userName: string;
-};
-
-type EditIssueDescriptionInput = ToolInput & {
-	issueId: string;
-	description: string;
-};
-
-type EditIssueTitleInput = ToolInput & {
-	issueId: string;
-	title: string;
-};
-
-type AddIssueTagInput = ToolInput &
-	IssueTargets & {
-		tagName: string;
-	};
-
-type RemoveIssueTagInput = ToolInput & {
-	issueId: string;
-	tagId: string;
-};
-
-type AddIssueAssigneeInput = ToolInput & {
-	issueId: string;
-	assigneeId?: string;
-	// Resolves to the config userId, the same identity that authors events.
-	self?: boolean;
-	// Names a person with no contributor record yet; a near-miss spelling makes
-	// a second, near-identical contributor.
-	assigneeName?: string;
-	// Required to create somebody new from a name; without it an unmatched name
-	// is refused rather than minting a contributor.
-	createUnlinked?: boolean;
-};
-
-type RemoveIssueAssigneeInput = ToolInput & {
-	issueId: string;
-	assigneeId: string;
-};
-
-type AddIssueCommentInput = ToolInput & {
-	issueId: string;
-	body: string;
-};
-
-type DeleteIssueCommentInput = ToolInput & {
-	commentId: string;
-};
-
-type EditIssueCommentInput = ToolInput & {
-	commentId: string;
-	body: string;
-};
-
-type AddIssueAttachmentInput = ToolInput & {
-	issueId: string;
-	name?: string;
-	// Either the bytes inline, as the GUI's upload endpoint sends them, or a
-	// path for a caller that already has the image on disk — an agent holding
-	// a screenshot should not have to base64 it through its own context.
-	dataBase64?: string;
-	filePath?: string;
-};
-
-const readAttachmentBytes = (
-	input: AddIssueAttachmentInput,
-): Result<{data: Buffer; name: string}> => {
-	if (input.filePath) {
-		try {
-			return succeeded('Read attachment file', {
-				data: fs.readFileSync(input.filePath),
-				name: input.name ?? path.basename(input.filePath),
-			});
-		} catch (error) {
-			return failed(
-				`Unable to read ${input.filePath}: ${
-					error instanceof Error ? error.message : String(error)
-				}`,
-			);
-		}
-	}
-
-	if (!input.dataBase64) return failed('Provide either filePath or dataBase64');
-
-	return succeeded('Decoded attachment data', {
-		data: Buffer.from(input.dataBase64, 'base64'),
-		name: input.name ?? 'image',
-	});
-};
-
-type DeleteIssueAttachmentInput = ToolInput & {
-	attachmentId: string;
-};
-
-type GetAttachmentBlobInput = ToolInput & {
-	fileName: string;
-};
-
-const getAttachmentMaxKb = (): number => {
-	const settings = loadSettingsFromConfig();
-	if (isFail(settings)) return DEFAULT_ATTACHMENT_MAX_KB;
-
-	return settings.value.attachmentMaxKb ?? DEFAULT_ATTACHMENT_MAX_KB;
-};
-
-const resolveRepoRoot = (repoRoot?: string): Result<string> => {
-	const result = resolveClosestEpiqProjectRoot(repoRoot ?? process.cwd());
-	if (isFail(result)) return failed(result.message);
-
-	return succeeded('Resolved Epiq repo root', result.value);
-};
-
-// The state branch and its worktree, once this process has seen them in place.
-//
-// Both checks spawn a subprocess, about 160ms for the pair, and every read and
-// every write runs them before doing anything. What they guard against is a
-// first call: a clone that never synced has no local branch for the worktree to
-// attach to. Neither comes and goes under a running server.
-//
-// Revalidated by looking for the worktree's own directory, which is what
-// disappears when another process relocates it — cheap enough to do every time,
-// and the case those checks would otherwise be catching.
-let ensuredBranch: {repoRoot: string; stateBranchRoot: string} | null = null;
-
-const boot = async (
-	repoRoot?: string,
-	options?: {pull?: boolean},
-): Promise<Result<BootResult>> => {
-	const repoRootResult = resolveRepoRoot(repoRoot);
-	if (isFail(repoRootResult)) return repoRootResult;
-
-	const stateBranchRootResult = getStateBranchRoot({
-		repoRoot: repoRootResult.value,
-	});
-
-	if (isFail(stateBranchRootResult)) {
-		return failed(stateBranchRootResult.message);
-	}
-
-	// This project's branch. `getProjectFileContents()` mints a fresh descriptor
-	// carrying the default name, so any project that named its state branch
-	// something else was ignored.
-	const stateBranchResult = getStateBranch(repoRootResult.value);
-	if (isFail(stateBranchResult)) return failed(stateBranchResult.message);
-
-	const inPlace =
-		ensuredBranch !== null &&
-		ensuredBranch.repoRoot === repoRootResult.value &&
-		ensuredBranch.stateBranchRoot === stateBranchRootResult.value &&
-		fs.existsSync(getEpiqDirPath(stateBranchRootResult.value));
-
-	if (!inPlace) {
-		// A clone that has never synced has no local state branch for the
-		// worktree to attach to. Deliberately not the full bootstrap the sync
-		// path runs: that also repairs the branch's contents with a commit.
-		const localBranchResult = await ensureLocalStateBranch({
-			repoRoot: repoRootResult.value,
-			stateBranchName: stateBranchResult.value,
-		});
-
-		if (isFail(localBranchResult)) return failed(localBranchResult.message);
-
-		const ensureWorktreeResult = await ensureStateBranchWorktree({
-			repoRoot: repoRootResult.value,
-			stateBranchRoot: stateBranchRootResult.value,
-			stateBranchName: stateBranchResult.value,
-		});
-
-		if (isFail(ensureWorktreeResult)) {
-			return failed(ensureWorktreeResult.message);
-		}
-
-		ensuredBranch = {
-			repoRoot: repoRootResult.value,
-			stateBranchRoot: stateBranchRootResult.value,
-		};
-	}
-
-	// MCP tools are local-only by default; fetching remote state is explicit.
-	if (options?.pull ?? false) {
-		const pullResult = await execGit({
-			cwd: stateBranchRootResult.value,
-			args: ['pull', '--ff-only'],
-		});
-
-		if (isFail(pullResult)) {
-			logger.info(3, pullResult.message);
-		}
-	}
-
-	const booted = {
-		repoRoot: repoRootResult.value,
-		stateBranchRoot: stateBranchRootResult.value,
-	};
-
-	// Before the load, never after — see log-signature for why that order is
-	// the one that stays correct when another machine writes mid-read.
-	const signature = logSignature(stateBranchRootResult.value);
-
-	// Reading the board should not mean deriving it again. Every call came
-	// through here re-reading the whole log and replaying every event: on a
-	// twenty-person decade that is 169 MB and 960,000 events, about eight
-	// seconds, for a board that had not changed since the last request a moment
-	// earlier.
-	//
-	// Only while live. A checkout in the past is materialized by the time-travel
-	// path and would be thrown away by a boot; `bootStateFromEventLog` refuses
-	// that case too, and skipping must not quietly take its place.
-	if (
-		accountedSignature(stateBranchRootResult.value) === signature &&
-		isStateInitialized() &&
-		getTimeTravelStatus().mode === 'live'
-	) {
-		return succeeded('Booted Epiq state, unchanged', booted);
-	}
-
-	const eventsResult = loadMergedEventsWithUnreadable(
-		stateBranchRootResult.value,
-	);
-	if (isFail(eventsResult)) return failed(eventsResult.message);
-
-	const bootResult = bootStateFromEventLog(
-		eventsResult.value.events,
-		eventsResult.value.unreadable,
-	);
-	if (isFail(bootResult)) return failed(bootResult.message);
-
-	// Recorded after the boot, so a failed one is retried rather than remembered
-	// as done. From here on a write advances this rather than invalidating it:
-	// the process applied what it wrote, so it does not have to read it back.
-	accountFor(stateBranchRootResult.value, signature);
-
-	return succeeded('Booted Epiq state', booted);
-};
-
-const getActor = (): Result<Actor> => {
-	const actorResult = loadSettingsFromConfig();
-	if (isFail(actorResult)) return failed(actorResult.message);
-
-	if (!actorResult.value.userId) return failed('Unable to retrieve user id');
-	if (!actorResult.value.userName) {
-		return failed('Unable to retrieve user name');
-	}
-
-	return succeeded('Resolved actor', {
-		userId: actorResult.value.userId,
-		userName: actorResult.value.userName,
-	});
-};
-
-const getStateResult = () => {
-	const stateResult = getSafeState();
-	if (isFail(stateResult)) return failed(stateResult.message);
-
-	return stateResult;
-};
-
-const getIssueTags = (ticket: Ticket) =>
-	(ticket.props.tags ?? [])
-		.map(tag => nodeRepo.getTag(tag))
-		.filter(tag => tag != undefined)
-		.map(tag => ({
-			id: tag.id,
-			name: tag.name,
-			color: getStringColor(tag.name),
-		}));
-
-// A contributor node's name is written once at create.contributor and never
-// updated; the event log carries the current one.
-const getLatestNamesFromLog = (): Map<string, string> => {
-	const stateResult = getSafeState();
-	const eventLog = isFail(stateResult) ? [] : stateResult.value.eventLog ?? [];
-	const byId = new Map<string, string>();
-
-	for (const event of eventLog) {
-		if (event.userId && event.userName) byId.set(event.userId, event.userName);
-	}
-
-	return byId;
-};
-
-// Shared so that every surface offering or matching a contributor agrees on the
-// answer; disagreement mints duplicate ids for the same person.
-const mergeRegistryNames = (
-	logNames: Map<string, string>,
-	registry: Record<string, Contributor>,
-): Map<string, string> => {
-	const byId = new Map(logNames);
-
-	// The registry always wins. The log's copy survives only for an id the
-	// registry has never seen — an author on a board written before renames
-	// were events.
-	for (const contributor of Object.values(registry)) {
-		byId.set(contributor.id, contributor.name);
-	}
-
-	return byId;
-};
-
-// A comment's id is its ULID, so sorting the ids is log order.
-const getIssueComments = (issueId: string): ApiIssueComment[] =>
-	nodeRepo
-		.getCommentsByIssue(issueId)
-		.sort((a, b) => (a.id < b.id ? -1 : a.id > b.id ? 1 : 0))
-		.map(comment => ({
-			id: comment.id,
-			author:
-				nodeRepo.getContributor(comment.authorId)?.name ??
-				comment.authorName ??
-				'Unknown',
-			createdAt: ulidTimeMs(comment.id),
-			body: comment.md,
-		}));
-
-const getIssueAssignees = (ticket: Ticket) =>
-	(ticket.props.assignees ?? [])
-		.map(assignee => nodeRepo.getContributor(assignee))
-		.filter(contributor => contributor != undefined)
-		.map(
-			({id, name}) =>
-				({
-					id,
-					name,
-					color: getStringColor(name),
-				} satisfies ApiIssue['assignees'][number]),
-		);
-
-const targetIds = (input: IssueTargets): Result<string[]> => {
-	const ids = [
-		...new Set([
-			...(input.issueIds ?? []),
-			...(input.issueId ? [input.issueId] : []),
-		]),
-	];
-
-	return ids.length
-		? succeeded('Resolved targets', ids)
-		: failed('Provide issueId or issueIds');
-};
-
-const batchResult = (
-	verb: string,
-	outcome: ApiBatchOutcome,
-): Result<ApiBatchOutcome> => {
-	const done = outcome.done.length;
-	const summary =
-		`${verb} ${done} issue${done === 1 ? '' : 's'}` +
-		(outcome.failed.length ? `, ${outcome.failed.length} failed` : '');
-
-	return done
-		? succeeded(summary, outcome)
-		: failed(
-				`${summary}: ${outcome.failed
-					.map(f => `${f.ref} ${f.reason}`)
-					.join('; ')}`,
-		  );
-};
-
-// Runs one write per target and sorts the outcomes. Each write persists on
-// its own: a rank is resolved against the state the write before it left, so
-// the saving is the round trip and the result in the caller's context, not
-// the persist.
-const forEachTarget = (
-	ids: string[],
-	write: (id: string) => Result<unknown>,
-): ApiBatchOutcome => {
-	const outcome: ApiBatchOutcome = {done: [], failed: []};
-
-	for (const id of ids) {
-		const result = write(id);
-		if (isFail(result)) {
-			outcome.failed.push({id, ref: nodeRef(id), reason: result.message});
-		} else {
-			outcome.done.push({id, ref: nodeRef(id)});
-		}
-	}
-
-	return outcome;
-};
-
-const findWritableIssue = (id: string): Result<Ticket> => {
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[id];
-	if (!issue || issue.isDeleted) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Target must be an issue');
-	if (issue.readonly) return failed('Issue is readonly');
-
-	return succeeded('Found issue', issue);
-};
-
-export const listBoards = async (input: ToolInput = {}) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const boards = Object.values(stateResult.value.nodes)
-		.filter(n => n.context === 'BOARD' && !n.isDeleted)
-		.map(n => ({
-			id: n.id,
-			ref: nodeRef(n.id),
-			title: n.title,
-			parentId: n.parentNodeId,
-			readonly: Boolean(n.readonly),
-		}));
-
-	return succeeded('Listed boards', boards);
-};
-
-export const listSwimlanes = async (input: ListSwimlanesInput = {}) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const swimlanes = Object.values(stateResult.value.nodes)
-		.filter(n => n.context === 'SWIMLANE' && !n.isDeleted)
-		.filter(n => !input.boardId || n.parentNodeId === input.boardId)
-		.map(n => ({
-			id: n.id,
-			title: n.title,
-			boardId: n.parentNodeId,
-			isClosed: n.id === CLOSED_SWIMLANE_ID,
-			readonly: Boolean(n.readonly),
-		}));
-
-	return succeeded('Listed swimlanes', swimlanes);
-};
-
-type GetIssueInput = ToolInput & {
-	idOrRef: string;
-};
-
-/**
- * One issue, by full id or by the 7-character ref the commit convention uses.
- *
- * The inverse of that convention had no tool: going from a ref back to a
- * ticket meant listing the whole board and filtering, and on a board of any
- * age that response is large enough to be unusable for reading one field.
- */
-export const getIssue = async (input: GetIssueInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const query = input.idOrRef.trim();
-	if (!query) return failed('Provide an issue id or ref');
-
-	const tickets = Object.values(stateResult.value.nodes)
-		.filter(isTicketNode)
-		.filter(n => !n.isDeleted);
-
-	const exact = tickets.find(n => n.id === query);
-
-	// `nodeRefMatches` is a substring match, so a short query can legitimately
-	// hit several tickets. Naming them beats guessing.
-	const matches = exact
-		? [exact]
-		: tickets.filter(n => nodeRefMatches(n.id, query));
-
-	if (matches.length === 0) {
-		return failed(`No issue matches "${query}"`);
-	}
-
-	if (matches.length > 1) {
-		return failed(
-			`"${query}" matches ${matches.length} issues: ${matches
-				.map(n => `${nodeRef(n.id)} (${sanitizeInlineText(n.title)})`)
-				.join(', ')}. Use a full ref or id.`,
-		);
-	}
-
-	const issue = matches[0] as Ticket;
-
-	return succeeded('Found issue', {
-		id: issue.id,
-		ref: nodeRef(issue.id),
-		title: sanitizeInlineText(issue.title),
-		description: issue.props.description ?? '',
-		createdAt: ulidTimeMs(issue.id),
-		parentNodeId: issue.parentNodeId!,
-		isClosed: issue.parentNodeId === CLOSED_SWIMLANE_ID,
-		readonly: Boolean(issue.readonly),
-		tags: getIssueTags(issue),
-		assignees: getIssueAssignees(issue),
-		comments: getIssueComments(issue.id),
-	} satisfies ApiIssueDetail);
-};
-
-const sameName = (a: string, b: string) =>
-	a.trim().toLowerCase() === b.trim().toLowerCase();
-
-export function listIssues(
-	input: ListIssuesInput & {brief: true},
-): Promise<Result<ApiIssueBrief[]>>;
-export function listIssues(input: ListIssuesInput): Promise<Result<ApiIssue[]>>;
-export async function listIssues(
-	input: ListIssuesInput,
-): Promise<Result<ApiIssue[] | ApiIssueBrief[]>> {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const nodes = stateResult.value.nodes;
-	const query = input.query?.trim().toLowerCase();
-
-	const tickets = Object.values(nodes)
-		.filter(isTicketNode)
-		.filter(n => !n.isDeleted)
-		.filter(n => input.includeClosed || n.parentNodeId !== CLOSED_SWIMLANE_ID)
-		.filter(n => {
-			if (!input.boardId) return true;
-			const swimlane = n.parentNodeId ? nodes[n.parentNodeId] : undefined;
-			return swimlane?.parentNodeId === input.boardId;
-		})
-		.filter(n => !input.swimlaneId || n.parentNodeId === input.swimlaneId)
-		.filter(
-			n =>
-				!input.tag ||
-				getIssueTags(n).some(tag => sameName(tag.name, input.tag!)),
-		)
-		.filter(
-			n =>
-				!input.assignee ||
-				getIssueAssignees(n).some(a => sameName(a.name, input.assignee!)),
-		)
-		.filter(
-			n =>
-				!query ||
-				n.title.toLowerCase().includes(query) ||
-				(n.props.description ?? '').toLowerCase().includes(query),
-		);
-
-	const swimlaneTitle = (n: Ticket) =>
-		sanitizeInlineText(
-			(n.parentNodeId ? nodes[n.parentNodeId]?.title : undefined) ?? '',
-		);
-
-	if (input.brief) {
-		return succeeded(
-			'Listed issues',
-			tickets.map(
-				n =>
-					({
-						id: n.id,
-						ref: nodeRef(n.id),
-						title: sanitizeInlineText(n.title),
-						swimlane: swimlaneTitle(n),
-						tags: getIssueTags(n).map(tag => tag.name),
-						assignees: getIssueAssignees(n).map(a => a.name),
-					} satisfies ApiIssueBrief),
-			),
-		);
-	}
-
-	return succeeded(
-		'Listed issues',
-		tickets.map(
-			n =>
-				({
-					id: n.id,
-					ref: nodeRef(n.id),
-					title: sanitizeInlineText(n.title),
-					description: n.props.description ?? '',
-					createdAt: ulidTimeMs(n.id),
-					parentNodeId: n.parentNodeId!,
-					isClosed: n.parentNodeId === CLOSED_SWIMLANE_ID,
-					readonly: Boolean(n.readonly),
-					tags: getIssueTags(n),
-					assignees: getIssueAssignees(n),
-				} satisfies ApiIssue),
-		),
-	);
-}
-
-export const createIssue = async (input: CreateIssueInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	// The other three title paths all sanitize; this one used to write the raw
-	// input, so a newline or a control character landed in the log for good.
-	const title = sanitizeInlineText(input.title);
-
-	if (!title) {
-		return failed('Issue title cannot be empty');
-	}
-
-	const description = input.description ?? '';
-
-	const overLong =
-		tooLong('Issue title', title, MAX_TITLE_LENGTH) ??
-		tooLong('Issue description', description, MAX_DESCRIPTION_LENGTH);
-	if (overLong) return failed(overLong);
-
-	if ((input.tagNames?.length ?? 0) > MAX_TAGS_PER_CREATE) {
-		return failed(`Cannot set more than ${MAX_TAGS_PER_CREATE} tags at once`);
-	}
-
-	if ((input.assigneeNames?.length ?? 0) > MAX_ASSIGNEES_PER_CREATE) {
-		return failed(
-			`Cannot set more than ${MAX_ASSIGNEES_PER_CREATE} assignees at once`,
-		);
-	}
-
-	const rankResult = resolveAndPersistRankForCreate(
-		input.parentId,
-		actorResult.value,
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(rankResult)) return rankResult;
-
-	const issueEventsResult = createIssueEvents({
-		name: title,
-		parent: input.parentId,
-		user: actorResult.value,
-		rank: rankResult.value,
-	});
-	if (isFail(issueEventsResult)) return issueEventsResult;
-
-	const events: AppEvent[] = [...issueEventsResult.value];
-	const issueId = events.find(e => e.action === 'add.issue')?.payload.id;
-	if (!issueId) return failed('Unable to determine created issue id');
-
-	if (description) {
-		events.push({
-			id: ulid(),
-			...actorResult.value,
-			action: 'edit.description',
-			payload: {id: issueId, md: description},
-		} satisfies AppEvent<'edit.description'>);
-	}
-
-	const tags: {id: string; name: string}[] = [];
-	for (const rawName of input.tagNames ?? []) {
-		const tagName = sanitizeInlineText(rawName).trim();
-		if (!tagName || tags.some(tag => tag.name === tagName)) continue;
-
-		const overLongTag = tooLong('Tag name', tagName, MAX_TAG_NAME_LENGTH);
-		if (overLongTag) return failed(overLongTag);
-
-		const existingTag = nodeRepo.findTagByName(tagName);
-		const tagId = existingTag?.id ?? ulid();
-
-		if (!existingTag) {
-			events.push({
-				id: ulid(),
-				...actorResult.value,
-				action: 'create.tag',
-				payload: {id: tagId, name: tagName},
-			} satisfies AppEvent<'create.tag'>);
-		}
-
-		events.push({
-			id: ulid(),
-			...actorResult.value,
-			action: 'add.issue.tag',
-			payload: {id: issueId, tag: tagId},
-		} satisfies AppEvent<'add.issue.tag'>);
-
-		tags.push({id: tagId, name: tagName});
-	}
-
-	const assignees: {id: string; name: string}[] = [];
-	for (const rawName of input.assigneeNames ?? []) {
-		const assigneeName = sanitizeInlineText(rawName).trim();
-		if (!assigneeName || assignees.some(a => a.name === assigneeName)) {
-			continue;
-		}
-
-		const overLongName = tooLong(
-			'Assignee name',
-			assigneeName,
-			MAX_ASSIGNEE_NAME_LENGTH,
-		);
-		if (overLongName) return failed(overLongName);
-
-		const existingAssignee = Object.values(stateResult.value.contributors).find(
-			contributor => contributor.name === assigneeName,
-		);
-		const assigneeId = existingAssignee?.id ?? ulid();
-
-		if (!existingAssignee) {
-			events.push({
-				id: ulid(),
-				...actorResult.value,
-				action: 'create.contributor',
-				payload: {id: assigneeId, name: assigneeName},
-			} satisfies AppEvent<'create.contributor'>);
-		}
-
-		events.push({
-			id: ulid(),
-			...actorResult.value,
-			action: 'add.issue.assignee',
-			payload: {id: issueId, assignee: assigneeId},
-		} satisfies AppEvent<'add.issue.assignee'>);
-
-		assignees.push({id: assigneeId, name: assigneeName});
-	}
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Created issue', {
-		id: issueId,
-		ref: nodeRef(issueId),
-		title,
-		parentId: input.parentId,
-		description,
-		tags,
-		assignees,
-	});
-};
-
-const closeOne = (id: string, actor: Actor, stateBranchRoot: string) => {
-	const issueResult = findWritableIssue(id);
-	if (isFail(issueResult)) return issueResult;
-
-	const rankResult = resolveAndPersistRankForMove(
-		CLOSED_SWIMLANE_ID,
-		id,
-		{at: 'end'},
-		actor,
-		stateBranchRoot,
-	);
-	if (isFail(rankResult)) return rankResult;
-
-	const event = {
-		id: ulid(),
-		...actor,
-		action: 'close.issue',
-		payload: {
-			id,
-			parent: CLOSED_SWIMLANE_ID,
-			rank: rankResult.value,
-		},
-	} satisfies AppEvent<'close.issue'>;
-
-	const results = materializeAndPersistAll([event], stateBranchRoot);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Closed issue', {id, ref: nodeRef(id)});
-};
-
-type IssueRef = {id: string; ref: string};
-
-export function closeIssue(
-	input: CloseIssueInput & {issueIds: string[]},
-): Promise<Result<ApiBatchOutcome>>;
-export function closeIssue(input: CloseIssueInput): Promise<Result<IssueRef>>;
-export async function closeIssue(
-	input: CloseIssueInput,
-): Promise<Result<ApiBatchOutcome | IssueRef>> {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const idsResult = targetIds(input);
-	if (isFail(idsResult)) return idsResult;
-
-	const close = (id: string) =>
-		closeOne(id, actorResult.value, bootResult.value.stateBranchRoot);
-
-	if (!input.issueIds) return close(idsResult.value[0]!);
-
-	return batchResult('Closed', forEachTarget(idsResult.value, close));
-}
-
-export const reopenIssue = async (input: ToolInput & {issueId: string}) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Target node is not issue');
-
-	if (issue.parentNodeId !== CLOSED_SWIMLANE_ID) {
-		return failed('Issue is not closed');
-	}
-
-	const previousParentId = resolveReopenParentFromLog(issue);
-
-	if (!previousParentId) {
-		return failed('Unable to resolve previous parent from issue history');
-	}
-
-	if (previousParentId === CLOSED_SWIMLANE_ID) {
-		return failed('Previous parent resolves to closed swimlane');
-	}
-
-	const previousParent = stateResult.value.nodes[previousParentId];
-
-	if (!previousParent) {
-		return failed('Previous parent no longer exists');
-	}
-
-	const rankResult = resolveAndPersistRankForMove(
-		previousParent.id,
-		issue.id,
-		{at: 'end'},
-		actorResult.value,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(rankResult)) return rankResult;
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'reopen.issue',
-		payload: {
-			id: issue.id,
-			parent: previousParent.id,
-			rank: rankResult.value,
-		},
-	} satisfies AppEvent<'reopen.issue'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Reopened issue', {
-		id: issue.id,
-		ref: nodeRef(issue.id),
-		parentId: previousParent.id,
-	});
-};
-
-const moveOne = (
-	id: string,
-	input: MoveIssueInput,
-	actor: Actor,
-	stateBranchRoot: string,
-) => {
-	const issueResult = findWritableIssue(id);
-	if (isFail(issueResult)) return issueResult;
-
-	const rankResult = resolveAndPersistRankForMove(
-		input.parentId,
-		id,
-		input.position ?? {at: 'end'},
-		actor,
-		stateBranchRoot,
-	);
-	if (isFail(rankResult)) return rankResult;
-
-	const event = {
-		id: ulid(),
-		...actor,
-		action: 'move.node',
-		payload: {
-			id,
-			parent: input.parentId,
-			rank: rankResult.value,
-		},
-	} satisfies AppEvent<'move.node'>;
-
-	const results = materializeAndPersistAll([event], stateBranchRoot);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Moved issue', {
-		id,
-		ref: nodeRef(id),
-		parentId: input.parentId,
-	});
-};
-
-export function moveIssue(
-	input: MoveIssueInput & {issueIds: string[]},
-): Promise<Result<ApiBatchOutcome>>;
-export function moveIssue(
-	input: MoveIssueInput,
-): Promise<Result<IssueRef & {parentId: string}>>;
-export async function moveIssue(
-	input: MoveIssueInput,
-): Promise<Result<ApiBatchOutcome | (IssueRef & {parentId: string})>> {
-	const repoRootResult = resolveRepoRoot(input.repoRoot);
-	if (isFail(repoRootResult)) return repoRootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateBranchRootResult = getStateBranchRoot({
-		repoRoot: repoRootResult.value,
-	});
-
-	if (isFail(stateBranchRootResult)) return stateBranchRootResult;
-
-	const eventsResult = loadMergedEventsWithUnreadable(
-		stateBranchRootResult.value,
-	);
-	if (isFail(eventsResult)) return eventsResult;
-
-	const bootStateResult = bootStateFromEventLog(
-		eventsResult.value.events,
-		eventsResult.value.unreadable,
-	);
-	if (isFail(bootStateResult)) return bootStateResult;
-
-	const idsResult = targetIds(input);
-	if (isFail(idsResult)) return idsResult;
-
-	const move = (id: string) =>
-		moveOne(id, input, actorResult.value, stateBranchRootResult.value);
-
-	if (!input.issueIds) return move(idsResult.value[0]!);
-
-	return batchResult('Moved', forEachTarget(idsResult.value, move));
-}
-
-export const createSwimlane = async (input: CreateSwimlaneInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const board = stateResult.value.nodes[input.boardId];
-	if (!board) return failed('Board not found');
-	if (!isBoardNode(board)) return failed('Target parent must be a board');
-	if (board.readonly)
-		return failed('Cannot add a swimlane to a readonly board');
-
-	// Boards carry no forced readonly of their own, so unlike the issue and
-	// swimlane mutations this one has to check the scrub itself. Without it a
-	// write lands on the state branch while the checkout is in the past.
-	if (getTimeTravelStatus().mode !== 'live') {
-		return failed('Cannot add a swimlane while time travelling');
-	}
-
-	const title = sanitizeInlineText(input.title);
-	if (!title.trim()) return failed('Swimlane title cannot be empty');
-
-	const overLongTitle = tooLong('Swimlane title', title, MAX_TITLE_LENGTH);
-	if (overLongTitle) return failed(overLongTitle);
-
-	const rankResult = resolveAndPersistRankForCreate(
-		input.boardId,
-		actorResult.value,
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(rankResult)) return rankResult;
-
-	const swimlaneId = ulid();
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'add.swimlane',
-		payload: {
-			id: swimlaneId,
-			name: title,
-			parent: input.boardId,
-			rank: rankResult.value,
-		},
-	} satisfies AppEvent<'add.swimlane'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Created swimlane', {
-		id: swimlaneId,
-		ref: nodeRef(swimlaneId),
-		title,
-		boardId: input.boardId,
-	});
-};
-
-export const editSwimlaneTitle = async (input: EditSwimlaneTitleInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const swimlane = stateResult.value.nodes[input.swimlaneId];
-
-	if (!swimlane) return failed('Swimlane not found');
-	if (!isSwimlaneNode(swimlane))
-		return failed('Edit target must be a swimlane');
-	if (swimlane.readonly) return failed('Cannot edit readonly swimlane');
-
-	const title = sanitizeInlineText(input.title);
-	if (!title.trim()) return failed('Swimlane title cannot be empty');
-
-	const overLongTitle = tooLong('Swimlane title', title, MAX_TITLE_LENGTH);
-	if (overLongTitle) return failed(overLongTitle);
-
-	if (swimlane.title === title) {
-		return succeeded('No changes made', {
-			id: input.swimlaneId,
-			ref: nodeRef(input.swimlaneId),
-			title,
-		});
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'edit.title',
-		payload: {
-			id: input.swimlaneId,
-			name: title,
-		},
-	} satisfies AppEvent<'edit.title'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Edited swimlane title', {
-		id: input.swimlaneId,
-		ref: nodeRef(input.swimlaneId),
-		title,
-	});
-};
-
-export const moveSwimlane = async (
-	input: MoveSwimlaneInput,
-): Promise<Result<{id: string; boardId: string}>> => {
-	const repoRootResult = resolveRepoRoot(input.repoRoot);
-	if (isFail(repoRootResult)) return repoRootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateBranchRootResult = getStateBranchRoot({
-		repoRoot: repoRootResult.value,
-	});
-
-	if (isFail(stateBranchRootResult)) return stateBranchRootResult;
-
-	const eventsResult = loadMergedEventsWithUnreadable(
-		stateBranchRootResult.value,
-	);
-	if (isFail(eventsResult)) return eventsResult;
-
-	const bootStateResult = bootStateFromEventLog(
-		eventsResult.value.events,
-		eventsResult.value.unreadable,
-	);
-	if (isFail(bootStateResult)) return bootStateResult;
-
-	const rankResult = resolveAndPersistRankForMove(
-		input.boardId,
-		input.swimlaneId,
-		input.position ?? {at: 'end'},
-		actorResult.value,
-		stateBranchRootResult.value,
-	);
-
-	if (isFail(rankResult)) return rankResult;
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'move.node',
-		payload: {
-			id: input.swimlaneId,
-			parent: input.boardId,
-			rank: rankResult.value,
-		},
-	} satisfies AppEvent<'move.node'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		stateBranchRootResult.value,
-	);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Moved swimlane', {
-		id: input.swimlaneId,
-		ref: nodeRef(input.swimlaneId),
-		boardId: input.boardId,
-	});
-};
-
-export const deleteSwimlane = async (input: DeleteSwimlaneInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const swimlane = stateResult.value.nodes[input.swimlaneId];
-
-	if (!swimlane) return failed('Swimlane not found');
-	if (!isSwimlaneNode(swimlane)) {
-		return failed('Delete target must be a swimlane');
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'delete.node',
-		payload: {
-			id: input.swimlaneId,
-		},
-	} satisfies AppEvent<'delete.node'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Deleted swimlane', {
-		id: input.swimlaneId,
-		ref: nodeRef(input.swimlaneId),
-	});
-};
-
-export const sync = async (input: SyncInput = {}) => {
-	setSyncing();
-	const repoRootResult = resolveRepoRoot(input.repoRoot);
-	if (isFail(repoRootResult)) return failed('Sync failed');
-
-	const actor = getActor();
-	if (isFail(actor)) return actor;
-
-	const result = await syncEpiqWithRemote({
-		cwd: repoRootResult.value,
-		ownEventFileName: getPersistFileName(actor.value),
-	});
-
-	if (isFail(result)) {
-		setSyncFailed(result.message);
-		return result;
-	}
-
-	setSynced();
-	return succeeded('Synced', result.value);
-};
-
-/**
- * A ticket's own event log, oldest first. Reads whatever is materialized rather
- * than booting, so it stays correct mid-scrub like the state beside it.
- */
-export const getIssueHistory = (
-	issueId: string,
-): Result<ApiIssueHistoryEntry[]> => {
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[issueId];
-	if (!issue) return failed('Issue not found');
-
-	return succeeded(
-		'Read issue history',
-		(issue.log ?? []).map(event => ({
-			id: event.id,
-			t: ulidTimeMs(event.id),
-			action: event.action,
-			label: describeEvent(event),
-			actor: {
-				id: event.userId,
-				name: event.userName,
-				color: getStringColor(event.userName),
-			},
-		})),
-	);
-};
-
-export const getEpiqState = async (input: ToolInput = {}) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	return succeeded('Retrieved Epiq state', {
-		root: bootResult.value.repoRoot,
-		stateBranchRoot: bootResult.value.stateBranchRoot,
-		nodes: stateResult.value.nodes,
-		rootNodeId: stateResult.value.rootNodeId,
-		contextNode: stateResult.value.contextNode,
-		selectedIndex: stateResult.value.selectedIndex,
-		eventLog: stateResult.value.eventLog,
-	});
-};
-
-// How far back the closed lane is sent. It is the one collection that grows
-// without bound — everything a team finishes lands there and stays — so a
-// decade of a twenty-person board is ninety-five thousand closed tickets
-// against a couple of hundred open ones.
-//
-// A year is what a reader is plausibly still looking for by eye. Reaching
-// further back is a request of its own rather than a cost every board pays on
-// every load.
-const CLOSED_TICKET_WINDOW_MS = 365 * 24 * 60 * 60 * 1000;
-
-// Reads whatever is currently materialized, live or a historical checkout. Must
-// never boot: booting discards an active time-travel checkout.
-export const deriveGuiState = (): Result<ApiState> => {
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const timeTravel = getTimeTravelStatus();
-	const forceReadonly = timeTravel.mode !== 'live';
-
-	const nodes = Object.values(stateResult.value.nodes);
-	const boards = nodes.filter(n => isBoardNode(n) && !n.isDeleted);
-
-	const closedSince = Date.now() - CLOSED_TICKET_WINDOW_MS;
-
-	const swimlanesByBoardId = new Map<string, Swimlane[]>();
-	const ticketsBySwimlaneId = new Map<string, Ticket[]>();
-
-	for (const node of nodes) {
-		if (node.isDeleted) continue;
-
-		if (isSwimlaneNode(node) && node.parentNodeId) {
-			const list = swimlanesByBoardId.get(node.parentNodeId) ?? [];
-			list.push(node);
-			swimlanesByBoardId.set(node.parentNodeId, list);
-		}
-
-		if (isTicketNode(node) && node.parentNodeId) {
-			// Everything a team ever finished lands in the closed lane and stays,
-			// so it is the one collection with no ceiling: a decade of a
-			// twenty-person board is ninety-five thousand tickets against a
-			// couple of hundred open ones, and sending them all is most of an
-			// 88 MB payload the browser has to parse before it can draw.
-			if (
-				node.parentNodeId === CLOSED_SWIMLANE_ID &&
-				ulidTimeMs(node.id) < closedSince
-			) {
-				continue;
-			}
-
-			const list = ticketsBySwimlaneId.get(node.parentNodeId) ?? [];
-			list.push(node);
-			ticketsBySwimlaneId.set(node.parentNodeId, list);
-		}
-	}
-
-	// The tickets the payload actually carries. Comments and attachments are
-	// built for these rather than for every ticket that has ever existed: a
-	// closed ticket nobody is sent does not need its conversation sent either.
-	const sentIssueIds = new Set<string>();
-
-	for (const tickets of ticketsBySwimlaneId.values()) {
-		for (const ticket of tickets) sentIssueIds.add(ticket.id);
-	}
-
-	const settingsRes = loadSettingsFromConfig();
-	if (isFail(settingsRes)) return settingsRes;
-
-	const commentsByIssueId: ApiState['commentsByIssueId'] = {};
-
-	// Grouped once rather than asked for per issue: asking per issue reads every
-	// comment on the board each time, which on a board with a hundred thousand
-	// of each never finishes.
-	const commentsByIssue = nodeRepo.getCommentsGroupedByIssue();
-
-	for (const issue of nodes.filter(isTicketNode)) {
-		if (issue.isDeleted || !sentIssueIds.has(issue.id)) continue;
-
-		commentsByIssueId[issue.id] = (commentsByIssue.get(issue.id) ?? []).map(
-			comment => {
-				const contributor = nodeRepo.getContributor(comment.authorId);
-
-				return {
-					id: comment.id,
-					issueId: comment.issue,
-					body: comment.md,
-					author: {
-						id: comment.authorId,
-						name: contributor?.name ?? 'Unknown',
-						// The author is optional in the log by design — the payload
-						// schema leaves it unconstrained so an event already written
-						// without one is not thrown away. So it can be missing here,
-						// and a colour is not worth taking the whole board down for.
-						color: getStringColor(
-							contributor?.name ?? comment.authorId ?? 'Unknown',
-						),
-					},
-					createdAt: ulidTimeMs(comment.id),
-				};
-			},
-		);
-	}
-
-	const attachmentOwners = new Map<string, string>();
-	for (const event of stateResult.value.eventLog) {
-		if (event.action === 'add.issue.attachment') {
-			attachmentOwners.set(event.payload.id, event.payload.author);
-		}
-	}
-
-	const attachmentsByIssueId: ApiState['attachmentsByIssueId'] = {};
-
-	const attachmentsByIssue = nodeRepo.getAttachmentsGroupedByIssue();
-
-	for (const issue of nodes.filter(isTicketNode)) {
-		if (issue.isDeleted || !sentIssueIds.has(issue.id)) continue;
-
-		attachmentsByIssueId[issue.id] = (
-			attachmentsByIssue.get(issue.id) ?? []
-		).map(attachment => ({
-			id: attachment.id,
-			issueId: attachment.issue,
-			name: attachment.name,
-			fileName: getAttachmentFileName(attachment.hash, attachment.ext),
-			bytes: attachment.bytes,
-			createdAt: ulidTimeMs(attachment.id),
-			canDelete:
-				attachmentOwners.get(attachment.id) === settingsRes.value.userId,
-		}));
-	}
-
-	return succeeded('Retrieved Epiq GUI state', {
-		boards: boards
-			.sort((a, b) => a.rank.localeCompare(b.rank))
-			.map(b => ({
-				id: b.id,
-				ref: nodeRef(b.id),
-				title: b.title,
-				readonly: Boolean(b.readonly) || forceReadonly,
-				swimlanes: (swimlanesByBoardId.get(b.id) ?? [])
-					.sort((a, b) => a.rank.localeCompare(b.rank))
-					.map(
-						swimlane =>
-							({
-								id: swimlane.id,
-								title: swimlane.title,
-								readonly: Boolean(swimlane.readonly) || forceReadonly,
-								issues: (ticketsBySwimlaneId.get(swimlane.id) ?? [])
-									.sort((a, b) => a.rank.localeCompare(b.rank))
-									.map(issue => ({
-										id: issue.id,
-										ref: nodeRef(issue.id),
-										title: sanitizeInlineText(issue.title),
-										description: issue.props.description ?? '',
-										createdAt: ulidTimeMs(issue.id),
-										readonly: Boolean(issue.readonly) || forceReadonly,
-										tags: getIssueTags(issue),
-										assignees: getIssueAssignees(issue),
-										parentNodeId: issue.parentNodeId!,
-										isClosed: issue.parentNodeId === CLOSED_SWIMLANE_ID,
-									})),
-								parentNodeId: swimlane.parentNodeId!,
-							} satisfies ApiSwimlane),
-					),
-			})),
-		tags: nodeRepo.getTags().map(x => ({
-			...x,
-			color: getStringColor(x.name),
-		})),
-		contributors: Object.values(stateResult.value.contributors).map(x => ({
-			...x,
-			color: getStringColor(x.name),
-		})),
-		user: {
-			name: settingsRes.value.userName ?? '',
-			id: settingsRes.value.userId ?? '',
-			color: getStringColor(settingsRes.value.userName ?? ''),
-		},
-		commentsByIssueId,
-		attachmentsByIssueId,
-		attachmentMaxKb: getAttachmentMaxKb(),
-		timeTravel,
-	} satisfies ApiState);
-};
-
-// The GUI asks for state after every mutation; the registry only needs to hear
-// about a project once per session.
-let lastRememberedRoot: string | null = null;
-
-const rememberOpenedProject = (repoRoot: string): void => {
-	if (repoRoot === lastRememberedRoot) return;
-
-	const result = recordRecentProject({root: repoRoot});
-	if (isFail(result)) {
-		logger.info(result.message);
-		return;
-	}
-
-	lastRememberedRoot = repoRoot;
-};
-
-export const getGuiState = async (
-	input: ToolInput = {},
-): Promise<Result<ApiState>> => {
-	// Fast path; the guarantee that a checkout survives a boot lives in
-	// `bootStateFromEventLog`, not here.
-	if (getTimeTravelStatus().mode !== 'live') {
-		return deriveGuiState();
-	}
-
-	// Never pull on a read: it would hang sandboxed or offline setups. Freshness
-	// is an explicit periodic sync's job.
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	rememberOpenedProject(bootResult.value.repoRoot);
-
-	return deriveGuiState();
-};
-
-export const editIssueDescription = async (
-	input: EditIssueDescriptionInput,
-) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Edit target must be an issue');
-	if (issue.readonly) return failed('Cannot edit readonly issue');
-
-	const overLongDescription = tooLong(
-		'Issue description',
-		input.description,
-		MAX_DESCRIPTION_LENGTH,
-	);
-	if (overLongDescription) return failed(overLongDescription);
-
-	const currentDescription = issue.props.description ?? '';
-
-	if (currentDescription === input.description) {
-		return succeeded('No changes made', {
-			id: input.issueId,
-			ref: nodeRef(input.issueId),
-			description: currentDescription,
-		});
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'edit.description',
-		payload: {
-			id: input.issueId,
-			md: input.description,
-		},
-	} satisfies AppEvent<'edit.description'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Edited issue description', {
-		id: input.issueId,
-		ref: nodeRef(input.issueId),
-		description: input.description,
-	});
-};
-
-export const editIssueTitle = async (input: EditIssueTitleInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Edit target must be an issue');
-	if (issue.readonly) return failed('Cannot edit readonly issue');
-
-	const title = sanitizeInlineText(input.title);
-
-	if (!title.trim()) {
-		return failed('Issue title cannot be empty');
-	}
-
-	const overLongTitle = tooLong('Issue title', title, MAX_TITLE_LENGTH);
-	if (overLongTitle) return failed(overLongTitle);
-
-	if (issue.title === title) {
-		return succeeded('No changes made', {
-			id: input.issueId,
-			ref: nodeRef(input.issueId),
-			title,
-		});
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'edit.title',
-		payload: {
-			id: input.issueId,
-			name: title,
-		},
-	} satisfies AppEvent<'edit.title'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Edited issue title', {
-		id: input.issueId,
-		ref: nodeRef(input.issueId),
-		title,
-	});
-};
-
-type TagRef = {tag: {id: string; name: string}};
-
-export function addIssueTag(
-	input: AddIssueTagInput & {issueIds: string[]},
-): Promise<Result<ApiBatchOutcome & TagRef>>;
-export function addIssueTag(
-	input: AddIssueTagInput,
-): Promise<Result<IssueRef & TagRef>>;
-export async function addIssueTag(
-	input: AddIssueTagInput,
-): Promise<Result<(ApiBatchOutcome & TagRef) | (IssueRef & TagRef)>> {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const idsResult = targetIds(input);
-	if (isFail(idsResult)) return idsResult;
-
-	const tagName = sanitizeInlineText(input.tagName).trim();
-	if (!tagName) return failed('Tag name cannot be empty');
-
-	const overLongTag = tooLong('Tag name', tagName, MAX_TAG_NAME_LENGTH);
-	if (overLongTag) return failed(overLongTag);
-
-	// Tagging needs no rank, so every target that checks out goes into one
-	// persist, behind the tag's creation if it is new.
-	const outcome: ApiBatchOutcome = {done: [], failed: []};
-
-	for (const id of idsResult.value) {
-		const issueResult = findWritableIssue(id);
-		if (isFail(issueResult)) {
-			outcome.failed.push({id, ref: nodeRef(id), reason: issueResult.message});
-		} else {
-			outcome.done.push({id, ref: nodeRef(id)});
-		}
-	}
-
-	if (!input.issueIds && outcome.failed[0]) {
-		return failed(outcome.failed[0].reason);
-	}
-
-	if (outcome.done.length === 0) {
-		return failed(batchResult('Tagged', outcome).message);
-	}
-
-	const existingTag = nodeRepo.findTagByName(tagName);
-
-	const tagId = existingTag?.id ?? ulid();
-
-	const events = [
-		...(existingTag
-			? []
-			: [
-					{
-						id: ulid(),
-						...actorResult.value,
-						action: 'create.tag',
-						payload: {
-							id: tagId,
-							name: tagName,
-						},
-					} satisfies AppEvent<'create.tag'>,
-			  ]),
-		...outcome.done.map(
-			({id}) =>
-				({
-					id: ulid(),
-					...actorResult.value,
-					action: 'add.issue.tag',
-					payload: {
-						id,
-						tag: tagId,
-					},
-				} satisfies AppEvent<'add.issue.tag'>),
-		),
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	const tag = {id: tagId, name: tagName};
-
-	if (!input.issueIds) {
-		return succeeded('Added issue tag', {...outcome.done[0]!, tag});
-	}
-
-	const batch = batchResult('Tagged', outcome);
-	return isFail(batch)
-		? batch
-		: succeeded(batch.message, {...batch.value, tag});
-}
-
-// Tombstone, not deletion: the id and every ticket reference survive in the
-// log; the tag just stops rendering anywhere, and its name is free again.
-export const tombstoneTag = async (
-	input: ToolInput & {tagId: string},
-): Promise<Result<{id: string; name: string}>> => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const tag = stateResult.value.tags[input.tagId];
-	if (!tag) return failed('Tag not found');
-	if (tag.tombstoned) return failed('Tag is already deleted');
-
-	const events = [
-		{
-			id: ulid(),
-			...actorResult.value,
-			action: 'tombstone.tag',
-			payload: {id: input.tagId},
-		} satisfies AppEvent<'tombstone.tag'>,
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Deleted tag', {id: tag.id, name: tag.name});
-};
-
-/**
- * Takes a board identity for the rest of this process. The actor is resolved
- * per write from the environment, so setting it here reaches the very next
- * event — no relaunch. A name given at launch wins: `applyActorNameArgument`
- * refuses to rename a process that was already told who it is.
- */
-export const assumeActor = async (
-	input: ToolInput & {name: string},
-): Promise<Result<{userId: string; userName: string; registered: boolean}>> => {
-	const applied = applyActorNameArgument(input.name, 'name');
-	if (isFail(applied)) return failed(applied.message);
-
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const actor = actorResult.value;
-
-	// Registered here rather than left to the first write, because the log file
-	// name is a lossy storage key — `claude/peter` sanitizes to `claude-peter`
-	// — and the registry is where display names are read from.
-	const isRegistered = Boolean(stateResult.value.contributors[actor.userId]);
-
-	if (!isRegistered) {
-		const results = materializeAndPersistAll(
-			[
-				{
-					id: ulid(),
-					...actor,
-					action: 'create.contributor',
-					payload: {id: actor.userId, name: actor.userName},
-				} satisfies AppEvent<'create.contributor'>,
-			],
-			bootResult.value.stateBranchRoot,
-		);
-
-		if (isFail(results)) return failed(results.message);
-	}
-
-	return succeeded(`Assumed ${actor.userName}`, {
-		...actor,
-		registered: !isRegistered,
-	});
-};
-
-export const restoreTag = async (
-	input: ToolInput & {tagId: string},
-): Promise<Result<{id: string; name: string}>> => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const tag = stateResult.value.tags[input.tagId];
-	if (!tag) return failed('Tag not found');
-	if (!tag.tombstoned) return failed('Tag is not deleted');
-
-	// The name may have been taken by a fresh tag in the meantime; two live
-	// tags with one name would be indistinguishable everywhere they are picked.
-	if (nodeRepo.findTagByName(tag.name)) {
-		return failed(`Cannot restore: another tag is already named "${tag.name}"`);
-	}
-
-	const events = [
-		{
-			id: ulid(),
-			...actorResult.value,
-			action: 'restore.tag',
-			payload: {id: input.tagId, name: tag.name},
-		} satisfies AppEvent<'restore.tag'>,
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Restored tag', {id: tag.id, name: tag.name});
-};
-
-export const removeIssueTag = async (input: RemoveIssueTagInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Untag target must be an issue');
-	if (issue.readonly) return failed('Cannot untag readonly issue');
-
-	if (!stateResult.value.tags[input.tagId]) {
-		return failed('Tag not found');
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'remove.issue.tag',
-		payload: {
-			id: input.issueId,
-			tag: input.tagId,
-		},
-	} satisfies AppEvent<'remove.issue.tag'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Removed issue tag', {
-		id: input.issueId,
-		ref: nodeRef(input.issueId),
-		tagId: input.tagId,
-	});
-};
-
-// Reads actors, not decoded events: the actor comes off the file name, so an
-// event this build cannot decode still proves its author wrote something. Going
-// through the decoded log instead would let an unreadable schema version make
-// somebody look unauthored, and the guard below would clear their removal.
-const findEventLogAuthor = async (
-	stateBranchRoot: string,
-	userId: string,
-): Promise<{id: string; name: string} | undefined> => {
-	const actorsResult = loadEventActors(stateBranchRoot);
-	if (isFail(actorsResult)) return undefined;
-
-	let name: string | undefined;
-
-	// Last write wins: a display name changes over time, the id does not.
-	for (const actor of actorsResult.value) {
-		if (actor.userId === userId) name = actor.userName ?? name;
-	}
-
-	return name === undefined ? undefined : {id: userId, name};
-};
-
-export const addIssueAssignee = async (input: AddIssueAssigneeInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Assign target must be an issue');
-	if (issue.readonly) return failed('Cannot assign readonly issue');
-
-	const targetId = input.self ? actorResult.value.userId : input.assigneeId;
-
-	// An id names a specific person, so an unknown one is an error rather than
-	// an invitation to invent a contributor.
-	if (targetId) {
-		const registered = stateResult.value.contributors[targetId];
-
-		// A contributor node is only written when somebody is explicitly created
-		// or assigned, so an author who was never assigned is absent from the
-		// registry. Register them under the id they already author with.
-		const authored = registered
-			? undefined
-			: await findEventLogAuthor(bootResult.value.stateBranchRoot, targetId);
-
-		if (!registered && !authored) return failed('Unknown assignee id');
-
-		const assignee = registered ?? authored!;
-
-		const events = [
-			...(registered
-				? []
-				: [
-						{
-							id: ulid(),
-							...actorResult.value,
-							action: 'create.contributor',
-							payload: {id: assignee.id, name: assignee.name},
-						} satisfies AppEvent<'create.contributor'>,
-				  ]),
-			{
-				id: ulid(),
-				...actorResult.value,
-				action: 'add.issue.assignee',
-				payload: {id: input.issueId, assignee: assignee.id},
-			} satisfies AppEvent<'add.issue.assignee'>,
-		];
-
-		const assignResults = materializeAndPersistAll(
-			events,
-			bootResult.value.stateBranchRoot,
-		);
-
-		if (isFail(assignResults)) return failed(assignResults.message);
-
-		return succeeded('Added issue assignee', {
-			id: input.issueId,
-			ref: nodeRef(input.issueId),
-			assignee: {id: assignee.id, name: assignee.name},
-		});
-	}
-
-	const assigneeName = sanitizeInlineText(input.assigneeName ?? '').trim();
-	if (!assigneeName) return failed('Provide assigneeId, self or assigneeName');
-
-	const overLongName = tooLong(
-		'Assignee name',
-		assigneeName,
-		MAX_ASSIGNEE_NAME_LENGTH,
-	);
-	if (overLongName) return failed(overLongName);
-
-	// Registry *and* event log, so this matches the same union a picker offers;
-	// the registry alone reports log-only authors as unknown.
-	const candidates = mergeRegistryNames(
-		getLatestNamesFromLog(),
-		stateResult.value.contributors,
-	);
-
-	const matches = [...candidates.entries()].filter(
-		([, candidateName]) => candidateName === assigneeName,
-	);
-
-	// Two people can share a display name; picking one silently assigns the
-	// wrong person half the time.
-	if (matches.length > 1) {
-		return failed(
-			`"${assigneeName}" matches ${matches.length} contributors (${matches
-				.map(([id]) => id)
-				.join(', ')}). Assign by assigneeId to choose.`,
-		);
-	}
-
-	const match = matches[0];
-
-	if (!match && !input.createUnlinked) {
-		return failed(
-			`No contributor named "${assigneeName}". Assign by id, or pass createUnlinked to add them as an external assignee.`,
-		);
-	}
-
-	const [assigneeId = ulid(), resolvedName = assigneeName] = match ?? [];
-
-	// Keyed on the registry, not on whether a name matched: somebody found only
-	// in the event log still needs a contributor record.
-	const isRegistered = Boolean(stateResult.value.contributors[assigneeId]);
-
-	const events = [
-		...(isRegistered
-			? []
-			: [
-					{
-						id: ulid(),
-						...actorResult.value,
-						action: 'create.contributor',
-						payload: {
-							id: assigneeId,
-							name: resolvedName,
-						},
-					} satisfies AppEvent<'create.contributor'>,
-			  ]),
-		{
-			id: ulid(),
-			...actorResult.value,
-			action: 'add.issue.assignee',
-			payload: {
-				id: input.issueId,
-				assignee: assigneeId,
-			},
-		} satisfies AppEvent<'add.issue.assignee'>,
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Added issue assignee', {
-		id: input.issueId,
-		ref: nodeRef(input.issueId),
-		assignee: {id: assigneeId, name: resolvedName},
-	});
-};
-
-// Tombstone, not deletion: the id and every reference to it survive, so only
-// the display name stops rendering. Refused for anyone who has authored an
-// event, since the log names them throughout and is never rewritten.
-export const tombstoneContributor = async (
-	input: ToolInput & {contributorId: string},
-): Promise<Result<{id: string; name: string}>> => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const contributor = stateResult.value.contributors[input.contributorId];
-	if (!contributor) return failed('Contributor not found');
-
-	const authored = await findEventLogAuthor(
-		bootResult.value.stateBranchRoot,
-		input.contributorId,
-	);
-
-	if (authored) {
-		return failed(
-			'Cannot remove a contributor who has authored events — their name appears throughout the log',
-		);
-	}
-
-	const events = [
-		{
-			id: ulid(),
-			...actorResult.value,
-			action: 'tombstone.contributor',
-			payload: {id: input.contributorId},
-		} satisfies AppEvent<'tombstone.contributor'>,
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Tombstoned contributor', {
-		id: input.contributorId,
-		name: REMOVED_CONTRIBUTOR_NAME,
-	});
-};
-
-// Removal never rewrites the log, so the `create.contributor` payload still
-// carries the original name. Last write wins.
-const findCreatedContributorName = async (
-	stateBranchRoot: string,
-	contributorId: string,
-): Promise<string | undefined> => {
-	const eventsResult = loadMergedEvents(stateBranchRoot);
-	if (isFail(eventsResult)) return undefined;
-
-	let name: string | undefined;
-
-	for (const event of eventsResult.value) {
-		if (event.action !== 'create.contributor') continue;
-
-		const payload = event.payload as {id?: string; name?: string};
-		if (payload.id === contributorId && payload.name) name = payload.name;
-	}
-
-	return name;
-};
-
-// The removal guard only sees the log this machine has pulled, so somebody
-// whose events have not arrived yet can be cleared by mistake. This makes that
-// reversible rather than forcing a network round-trip inside a read path.
-export const restoreContributor = async (
-	input: ToolInput & {contributorId: string},
-): Promise<Result<{id: string; name: string}>> => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const contributor = stateResult.value.contributors[input.contributorId];
-	if (!contributor) return failed('Contributor not found');
-
-	if (!contributor.tombstoned) {
-		return failed('Contributor is not tombstoned');
-	}
-
-	const originalName = await findCreatedContributorName(
-		bootResult.value.stateBranchRoot,
-		input.contributorId,
-	);
-
-	if (!originalName) {
-		return failed(
-			'Cannot restore this contributor — no original name found in the event log',
-		);
-	}
-
-	const events = [
-		{
-			id: ulid(),
-			...actorResult.value,
-			action: 'restore.contributor',
-			payload: {id: input.contributorId, name: originalName},
-		} satisfies AppEvent<'restore.contributor'>,
-	];
-
-	const results = materializeAndPersistAll(
-		events,
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Restored contributor', {
-		id: input.contributorId,
-		name: originalName,
-	});
-};
-
-export const getBoardContributors = async (
-	input: ToolInput & {boardId?: string} = {},
-): Promise<
-	Result<
-		(ApiAssignee & {
-			isSelf: boolean;
-			isExternal: boolean;
-			isRemoved: boolean;
-			hasAuthoredAnywhere: boolean;
-		})[]
-	>
-> => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const eventsResult = loadMergedEvents(bootResult.value.stateBranchRoot);
-	if (isFail(eventsResult)) return failed(eventsResult.message);
-
-	const scopedEvents = input.boardId
-		? filterEventsForBoard(eventsResult.value, input.boardId)
-		: eventsResult.value;
-
-	// Last write wins: events arrive in chronological order.
-	const byId = new Map<string, string>();
-	// Board-scoped, and kept apart from the merged map so the union below cannot
-	// erase "has actually worked on this board".
-	const authorIds = new Set<string>();
-
-	// Unfiltered, unlike `authorIds`: removal is refused for anyone who has
-	// authored anywhere, so this must read the same actors that guard does —
-	// off the file names, including events this build cannot decode.
-	const workspaceAuthorIds = new Set<string>();
-
-	const actorsResult = loadEventActors(bootResult.value.stateBranchRoot);
-	if (isFail(actorsResult)) return failed(actorsResult.message);
-
-	for (const actor of actorsResult.value) {
-		if (actor.userId) workspaceAuthorIds.add(actor.userId);
-	}
-
-	for (const event of scopedEvents) {
-		if (!event.userId) continue;
-
-		byId.set(event.userId, event.userName ?? '');
-		authorIds.add(event.userId);
-	}
-
-	const registry = stateResult.value.contributors;
-	const namesById = mergeRegistryNames(byId, registry);
-
-	const contributors = [...namesById.entries()].map(([id, name]) => ({
-		id,
-		name,
-		color: getStringColor(name),
-		isSelf: id === actorResult.value.userId,
-		// Board-scoped: means "has not worked on this board", not "is not in the
-		// history". Never stored, so it self-corrects.
-		isExternal: !authorIds.has(id),
-		// Read off the record, not compared against the placeholder name, so
-		// somebody genuinely called "removed" is not reported as already removed.
-		isRemoved: registry[id]?.tombstoned === true,
-		// Workspace-wide: their name is in the log, which is what blocks removal.
-		hasAuthoredAnywhere: workspaceAuthorIds.has(id),
-	}));
-
-	return succeeded('Listed board contributors', contributors);
-};
-
-export const removeIssueAssignee = async (input: RemoveIssueAssigneeInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Unassign target must be an issue');
-	if (issue.readonly) return failed('Cannot unassign readonly issue');
-
-	if (!stateResult.value.contributors[input.assigneeId]) {
-		return failed('Assignee not found');
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'remove.issue.assignee',
-		payload: {
-			id: input.issueId,
-			assignee: input.assigneeId,
-		},
-	} satisfies AppEvent<'remove.issue.assignee'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Removed issue assignee', {
-		id: input.issueId,
-		ref: nodeRef(input.issueId),
-		assigneeId: input.assigneeId,
-	});
-};
-
-export const addIssueComment = async (input: AddIssueCommentInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Comment target must be an issue');
-	if (issue.readonly) return failed('Cannot comment on readonly issue');
-
-	const body = input.body.trim();
-
-	if (!body) {
-		return failed('Comment cannot be empty');
-	}
-
-	// The single gate for MCP and both GUI transports.
-	if (body.length > MAX_COMMENT_LENGTH) {
-		return failed(
-			`Comment cannot exceed ${MAX_COMMENT_LENGTH} characters (got ${body.length})`,
-		);
-	}
-
-	const commentId = ulid();
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'add.issue.comment',
-		payload: {
-			id: commentId,
-			issue: input.issueId,
-			md: body,
-			author: actorResult.value.userId,
-		},
-	} satisfies AppEvent<'add.issue.comment'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Added issue comment', {
-		id: commentId,
-		issueId: input.issueId,
-		body,
-	});
-};
-
-export const deleteIssueComment = async (input: DeleteIssueCommentInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const commentEvent = stateResult.value.eventLog.find(
-		(event): event is AppEvent<'add.issue.comment'> =>
-			event.action === 'add.issue.comment' &&
-			event.payload.id === input.commentId,
-	);
-
-	if (!commentEvent) {
-		return failed('Unable to resolve comment');
-	}
-
-	if (commentEvent.payload.author !== actorResult.value.userId) {
-		return failed('You can only delete your own comments');
-	}
-
-	const issue = stateResult.value.nodes[commentEvent.payload.issue];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Comment target must be an issue');
-	if (issue.readonly) return failed('Cannot delete comment on readonly issue');
-
-	const alreadyDeleted = stateResult.value.eventLog.some(
-		event =>
-			event.action === 'delete.issue.comment' &&
-			event.payload.id === input.commentId,
-	);
-
-	if (alreadyDeleted) {
-		return succeeded('Comment already deleted', {
-			id: input.commentId,
-			issueId: commentEvent.payload.issue,
-		});
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'delete.issue.comment',
-		payload: {
-			id: input.commentId,
-			issue: commentEvent.payload.issue,
-		},
-	} satisfies AppEvent<'delete.issue.comment'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Deleted issue comment', {
-		id: input.commentId,
-		issueId: commentEvent.payload.issue,
-	});
-};
-
-export const editIssueComment = async (input: EditIssueCommentInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const commentEvent = stateResult.value.eventLog.find(
-		(event): event is AppEvent<'add.issue.comment'> =>
-			event.action === 'add.issue.comment' &&
-			event.payload.id === input.commentId,
-	);
-
-	if (!commentEvent) {
-		return failed('Unable to resolve comment');
-	}
-
-	if (commentEvent.payload.author !== actorResult.value.userId) {
-		return failed('You can only edit your own comments');
-	}
-
-	const issue = stateResult.value.nodes[commentEvent.payload.issue];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Comment target must be an issue');
-	if (issue.readonly) return failed('Cannot edit comment on readonly issue');
-
-	const deleted = stateResult.value.eventLog.some(
-		event =>
-			event.action === 'delete.issue.comment' &&
-			event.payload.id === input.commentId,
-	);
-
-	if (deleted) return failed('Comment was deleted');
-
-	const body = input.body.trim();
-
-	if (!body) {
-		return failed('Comment cannot be empty');
-	}
-
-	if (body.length > MAX_COMMENT_LENGTH) {
-		return failed(
-			`Comment cannot exceed ${MAX_COMMENT_LENGTH} characters (got ${body.length})`,
-		);
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'edit.issue.comment',
-		payload: {
-			id: input.commentId,
-			issue: commentEvent.payload.issue,
-			md: body,
-		},
-	} satisfies AppEvent<'edit.issue.comment'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Edited issue comment', {
-		id: input.commentId,
-		issueId: commentEvent.payload.issue,
-		body,
-	});
-};
-
-export const addIssueAttachment = async (input: AddIssueAttachmentInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const issue = stateResult.value.nodes[input.issueId];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Attachment target must be an issue');
-	if (issue.readonly) return failed('Cannot attach to readonly issue');
-
-	const bytesResult = readAttachmentBytes(input);
-	if (isFail(bytesResult)) return bytesResult;
-
-	const written = writeAttachmentBlob(
-		bootResult.value.stateBranchRoot,
-		bytesResult.value.data,
-		getAttachmentMaxKb(),
-	);
-	if (isFail(written)) return written;
-
-	const name = sanitizeInlineText(bytesResult.value.name).trim() || 'image';
-	const attachmentId = ulid();
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'add.issue.attachment',
-		payload: {
-			id: attachmentId,
-			issue: input.issueId,
-			author: actorResult.value.userId,
-			hash: written.value.hash,
-			ext: written.value.ext,
-			name,
-			bytes: written.value.bytes,
-		},
-	} satisfies AppEvent<'add.issue.attachment'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	const fileName = getAttachmentFileName(written.value.hash, written.value.ext);
-
-	return succeeded('Added issue attachment', {
-		id: attachmentId,
-		issueId: input.issueId,
-		fileName,
-		bytes: written.value.bytes,
-		markdown: getAttachmentMarkdown(name, fileName),
-	});
-};
-
-export const deleteIssueAttachment = async (
-	input: DeleteIssueAttachmentInput,
-) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	const actorResult = getActor();
-	if (isFail(actorResult)) return actorResult;
-
-	const stateResult = getStateResult();
-	if (isFail(stateResult)) return stateResult;
-
-	const attachmentEvent = stateResult.value.eventLog.find(
-		(event): event is AppEvent<'add.issue.attachment'> =>
-			event.action === 'add.issue.attachment' &&
-			event.payload.id === input.attachmentId,
-	);
-
-	if (!attachmentEvent) {
-		return failed('Unable to resolve attachment');
-	}
-
-	if (attachmentEvent.payload.author !== actorResult.value.userId) {
-		return failed('You can only delete your own attachments');
-	}
-
-	const issue = stateResult.value.nodes[attachmentEvent.payload.issue];
-
-	if (!issue) return failed('Issue not found');
-	if (!isTicketNode(issue)) return failed('Attachment target must be an issue');
-	if (issue.readonly) {
-		return failed('Cannot delete attachment on readonly issue');
-	}
-
-	const alreadyDeleted = stateResult.value.eventLog.some(
-		event =>
-			event.action === 'delete.issue.attachment' &&
-			event.payload.id === input.attachmentId,
-	);
-
-	if (alreadyDeleted) {
-		return succeeded('Attachment already deleted', {
-			id: input.attachmentId,
-			issueId: attachmentEvent.payload.issue,
-		});
-	}
-
-	const event = {
-		id: ulid(),
-		...actorResult.value,
-		action: 'delete.issue.attachment',
-		payload: {
-			id: input.attachmentId,
-			issue: attachmentEvent.payload.issue,
-		},
-	} satisfies AppEvent<'delete.issue.attachment'>;
-
-	const results = materializeAndPersistAll(
-		[event],
-		bootResult.value.stateBranchRoot,
-	);
-
-	if (isFail(results)) return failed(results.message);
-
-	return succeeded('Deleted issue attachment', {
-		id: input.attachmentId,
-		issueId: attachmentEvent.payload.issue,
-	});
-};
-
-/**
- * Synced blobs are untrusted input; `resolveAttachmentBlob` does the name,
- * hash, and magic-byte validation.
- */
-export const getAttachmentBlob = async (input: GetAttachmentBlobInput) => {
-	const bootResult = await boot(input.repoRoot, {pull: false});
-	if (isFail(bootResult)) return bootResult;
-
-	return resolveAttachmentBlob(
-		bootResult.value.stateBranchRoot,
-		input.fileName,
-		getAttachmentMaxKb(),
-	);
-};
+export {sync} from './api/sync.js';
+export {
+	getIssue,
+	listIssues,
+	createIssue,
+	closeIssue,
+	reopenIssue,
+	moveIssue,
+	getIssueHistory,
+	editIssueDescription,
+	editIssueTitle,
+} from './api/issues.js';
+export {
+	listBoards,
+	listSwimlanes,
+	createSwimlane,
+	editSwimlaneTitle,
+	moveSwimlane,
+	deleteSwimlane,
+} from './api/boards.js';
+export {
+	addIssueTag,
+	tombstoneTag,
+	restoreTag,
+	removeIssueTag,
+} from './api/tags.js';
+export {
+	assumeActor,
+	addIssueAssignee,
+	tombstoneContributor,
+	restoreContributor,
+	getBoardContributors,
+	removeIssueAssignee,
+} from './api/contributors.js';
+export {
+	addIssueComment,
+	deleteIssueComment,
+	editIssueComment,
+} from './api/comments.js';
+export {
+	addIssueAttachment,
+	deleteIssueAttachment,
+	getAttachmentBlob,
+} from './api/attachments.js';
+export {getEpiqState, deriveGuiState, getGuiState} from './api/state.js';
